+++ apps/next/package.json
@@ -17,8 +17,6 @@
 	"dependencies": {
 		"@emiljohansson/random-string": "1.1.2",
 		"@faker-js/faker": "7.4.0",
-		"@preact/signals-core": "1.5.0",
-		"@preact/signals-react": "1.3.7",
 		"@radix-ui/react-accessible-icon": "1.0.2",
 		"@radix-ui/react-label": "1.0.0",
 		"@radix-ui/react-progress": "1.0.0",
