+++ packages/lib/package.json
@@ -9,6 +9,8 @@
 		"test:ci": "npm run test"
 	},
 	"dependencies": {
+		"@preact/signals-core": "1.5.0",
+		"@preact/signals-react": "1.3.7",
 		"next": "14.0.3",
 		"react": "18.2.0",
 		"react-dom": "18.2.0"
